Script que monta el backup diario

+++ montarBackup.php
@@ -12,33 +12,37 @@ chdir("/home/sdvpruebas/backups/scripts");
 
 include ("./funciones.php");
 
-$txtNombreArchivoTar = date("Ymd") . "-sipive";
-if( file_exists( DESTINO_BACKUP_BD . "/" . $txtNombreArchivoTar . ".tar.gz" )){
+$txtBaseDatosDestino = ( isset($argv[1] ) )? $argv[1] : "";
 
-    mensajeLog("Descomprimiendo backup " . $txtNombreArchivoTar );
-    $txtComando = "tar -zxvf " . DESTINO_BACKUP_BD . "/" . $txtNombreArchivoTar . ".tar.gz -C /";
-    mensajeLog($txtComando);
-    //exec($txtComando);
+if( $txtBaseDatosDestino != ""){
 
-    mensajeLog("Reemplaza " . NOMBRE_BD . " por sipive_capacitacion");
-    $txtComando = "sed -i 's/" . NOMBRE_BD . "/sipive_capacitacion/g' " .  DESTINO_BACKUP_BD . "/" . $txtNombreArchivoTar . ".sql";
-    mensajeLog($txtComando);
-    //exec($txtComando);
+	$txtNombreArchivoTar = date("Ymd") . "-sipive";
+	if( file_exists( DESTINO_BACKUP_BD . "/" . $txtNombreArchivoTar . ".tar.gz" )){
 
-    mensajeLog("Montando la base de datos");
-    $txtComando = "mysql -u" . USUARIO_BD . " -p" . CLAVE_BD . " sipive_capacitacion < " . DESTINO_BACKUP_BD . "/" . $txtNombreArchivoTar . ".sql";
-    mensajeLog($txtComando);
-    //exec($txtComando);
+    		mensajeLog("Descomprimiendo backup " . $txtNombreArchivoTar . ".tar.gz" );
+    		$txtComando = "tar -zxvf " . DESTINO_BACKUP_BD . "/" . $txtNombreArchivoTar . ".tar.gz -C /";
+    		//mensajeLog($txtComando);
+    		exec($txtComando);
+	
+    		mensajeLog("Reemplaza " . NOMBRE_BD . " por " . $txtBaseDatosDestino);
+    		$txtComando = "sed -i 's/`" . NOMBRE_BD . "/`" . $txtBaseDatosDestino . "/g' " .  DESTINO_BACKUP_BD . "/" . $txtNombreArchivoTar . ".sql";
+    		mensajeLog($txtComando);
+    		//exec($txtComando);
+	
+    		mensajeLog("Montando la base de datos");
+    		$txtComando = "mysql -u" . USUARIO_BD . " -p" . CLAVE_BD . " " . $txtBaseDatosDestino . " < " . DESTINO_BACKUP_BD . "/" . $txtNombreArchivoTar . ".sql";
+    		mensajeLog($txtComando);
+    		//exec($txtComando);
+	
+    		mensajeLog("Limpiando archivos");
+    		$txtComando = "rm -rf " . DESTINO_BACKUP_BD . "/" . $txtNombreArchivoTar . ".sql";
+    		mensajeLog($txtComando);
+    		//exec($txtComando);
 
-    mensajeLog("Limpiando archivos");
-    $txtComando = "rm -rf " . DESTINO_BACKUP_BD . "/" . $txtNombreArchivoTar . ".sql";
-    mensajeLog($txtComando);
-    //exec($txtComando);
-
+	}else{
+    		mensajeLog("No se encontró el backup " . $txtNombreArchivoTar);
+	}
 }else{
-    mensajeLog("No se encontró el backup " . $txtNombreArchivoTar);
+	mensajeLog("Debe definir la base de datos destino");
 }
-
-
-
 ?>
